+++ src/components/ProductForm/index.ts
@@ -1,0 +1,1 @@
+export { ProductForm as default, ProductForm } from './ProductForm';
